+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/block/TomatoCropBlock.java
@@ -21,7 +21,6 @@ import net.minecraft.world.level.block.CropBlock;
 import net.minecraft.world.level.block.FarmBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
 import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.level.pathfinder.PathComputationType;
@@ -55,7 +54,7 @@ public class TomatoCropBlock extends CropBlock {
 
     @Override
     public int getMaxAge() {
-        return 5;
+        return 7;
     }
 
     @SuppressWarnings("deprecation")
@@ -67,9 +66,9 @@ public class TomatoCropBlock extends CropBlock {
             return dropsOriginal;
         }
         return
-                state.getValue(AGE) == 5 ?  List.of(new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
-                        state.getValue(AGE) == 6 ? List.of(new ItemStack(ItemRegistry.TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
-                                state.getValue(AGE) == 7 ? List.of(new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
+                state.getValue(AGE) == 4 || state.getValue(AGE) == 5 || state.getValue(AGE) == 6 ?  List.of(new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
+                        state.getValue(AGE) == 7 ? List.of(new ItemStack(ItemRegistry.TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
+                                state.getValue(AGE) == 8 ? List.of(new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), randomTomatoes), new ItemStack(this, 2)) :
                                         List.of(new ItemStack(this, 1));
     }
 
@@ -98,19 +97,19 @@ public class TomatoCropBlock extends CropBlock {
     @Override
     public @NotNull InteractionResult use(BlockState state, @NotNull Level level, @NotNull BlockPos pos, @NotNull Player player, @NotNull InteractionHand interactionHand, @NotNull BlockHitResult blockHitResult) {
         int age = state.getValue(AGE);
-        boolean reachedTomatoAge = age == 5;
+        boolean reachedTomatoAge = age == 4;
         if (!reachedTomatoAge && player.getItemInHand(interactionHand).is(Items.BONE_MEAL)) {
             return InteractionResult.PASS;
-        } else if (age > 4) {
+        } else if (age > 3) {
             ItemStack stack = player.getItemInHand(interactionHand);
             if (stack.getItem() instanceof KnifeItem) {
                 int randomTomatoes = 3 + level.random.nextInt(4);
                 popResource(level, pos,
-                        age == 5 ?  new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), randomTomatoes) :
-                        age == 6 ? new ItemStack(ItemRegistry.TOMATO.get(), randomTomatoes) :
-                        age == 7 ? new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), randomTomatoes) : ItemStack.EMPTY);
+                        age == 4 || age == 5 || age == 6 ? new ItemStack(ItemRegistry.UNRIPE_TOMATO.get(), randomTomatoes) :
+                        age == 7 ? new ItemStack(ItemRegistry.TOMATO.get(), randomTomatoes) :
+                        age == 8 ? new ItemStack(ItemRegistry.ROTTEN_TOMATO.get(), randomTomatoes) : ItemStack.EMPTY);
                 level.playSound(null, pos, SoundEvents.ITEM_PICKUP, SoundSource.BLOCKS, 1.0F, 0.8F + level.random.nextFloat() * 0.4F);
-                BlockState newBlockState = state.setValue(AGE, 4);
+                BlockState newBlockState = state.setValue(AGE, 3);
                 level.setBlock(pos, newBlockState, 2);
                 level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(player, newBlockState));
                 stack.hurtAndBreak(1, player, (livingEntity) -> livingEntity.broadcastBreakEvent(interactionHand));
@@ -127,15 +126,15 @@ public class TomatoCropBlock extends CropBlock {
 
     @Override
     public boolean isRandomlyTicking(@NotNull BlockState state) {
-        return state.getValue(AGE) < 7;
+        return state.getValue(AGE) < 8;
     }
 
     @Override
     public void randomTick(@NotNull BlockState state, @NotNull ServerLevel $$1, @NotNull BlockPos $$2, @NotNull RandomSource $$3) {
         int age = state.getValue(AGE);
-        if (age < 7 && $$3.nextInt(5) == 0 && $$1.getRawBrightness($$2.above(), 0) >= 9) {
+        if (age < 8 && $$3.nextInt(5) == 0 && $$1.getRawBrightness($$2.above(), 0) >= 9) {
             BlockState $$5 = state;
-            if (age == 6) {
+            if (age == 7) {
                 if ($$3.nextInt(4) == 0) { // 25% chance
                     $$5 = state.setValue(AGE, age + 1);
                 }
@@ -154,20 +153,21 @@ public class TomatoCropBlock extends CropBlock {
 
     @Override
     public boolean isValidBonemealTarget(@NotNull LevelReader levelReader, @NotNull BlockPos pos, BlockState state, boolean $$3) {
-        return state.getValue(AGE) < 5;
+        return state.getValue(AGE) < 4;
     }
 
     static {
-        AGE = BlockStateProperties.AGE_7;
+        AGE = IntegerProperty.create("age", 0, 8);
         SHAPE_BY_AGE = new VoxelShape[]{
-                Block.box(0.0, 0.0, 0.0, 16.0, 2.0, 16.0),
                 Block.box(0.0, 0.0, 0.0, 16.0, 4.0, 16.0),
                 Block.box(0.0, 0.0, 0.0, 16.0, 8.0, 16.0),
-                Block.box(0.0, 0.0, 0.0, 16.0, 10.0, 16.0),
+                Block.box(0.0, 0.0, 0.0, 16.0, 11.0, 16.0),
                 Block.box(0.0, 0.0, 0.0, 16.0, 13.0, 16.0),
-                Block.box(0.0, 0.0, 0.0, 16.0, 13.0, 16.0),
-                Block.box(0.0, 0.0, 0.0, 16.0, 13.0, 16.0),
-                Block.box(0.0, 0.0, 0.0, 16.0, 13.0, 16.0)};
+                Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0),
+                Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0),
+                Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0),
+                Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0),
+                Block.box(0.0, 0.0, 0.0, 16.0, 14.0, 16.0)};
     }
 
 }