+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/data/BrutalItemTagsProvider.java
@@ -9,6 +9,7 @@ import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
 import net.minecraft.data.tags.TagsProvider;
 import net.minecraft.tags.ItemTags;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.common.data.ExistingFileHelper;
@@ -72,6 +73,8 @@ public class BrutalItemTagsProvider extends ItemTagsProvider {
         });
 
         tag(ItemTags.PIGLIN_LOVED).add(ItemRegistry.GOLDEN_KNIFE.get());
+
+        tag(TagRegistry.Items.BUCKETS_WATER).add(Items.WATER_BUCKET);
 
     }
 }