+++ Forge/changelog.md
@@ -1,2 +1,10 @@
+1.0.3:
+* Added Corn, Cucumber and Lettuce Seeds to villager house loot
+* Added recipe for Pickle
+* Increased max age on Tomatoes
+
+1.0.2:
+* Updated textures
+
 1.0.1:
 * Fixed Tomato projectile particles